+++ CHANGELOG.md
@@ -25,14 +25,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Class:
     - `OxidEsales\Codeception\Admin\CoreSetting\SettingsTab`
     - `OxidEsales\Codeception\Page\Account\MyDownloads`
+    - `OxidEsales\Codeception\Admin\Languages`
+    - `OxidEsales\Codeception\Admin\Tools`
 - Method:
     - `OxidEsales\Codeception\Admin\CoreSettings::openSettingsTab`
     - `OxidEsales\Codeception\Admin\Orders::find`
     - `OxidEsales\Codeception\Admin\Orders::openDownloadsTab`
     - `OxidEsales\Codeception\Admin\Products::openDownloadsTab`
     - `OxidEsales\Codeception\Page\Account\Component\AccountNavigation::openMyDownloadsPage`
-    - 
-    
+    - `OxidEsales\Codeception\Admin\Component\AdminMenu::openLanguages`
+    - `OxidEsales\Codeception\Admin\Component\AdminMenu::openTools`
+   
 ## [1.3.3] - 2020-07-06
 
 ### Added